Level 1 basic mechanic
The basic construction of the first level scene has been completed, and the attack animations of the characters and enemies have been completed. The characters we control have a health bar UI and a death animation. The basic first level has been completed. In the first level, I have only designed a basic trap, and I still have three or four traps that have not been added. The sprite of the Gnome enemy was difficult to find until I found an element on itcho and put it in. The entire game is only short of completion: the enemy lacks a health system and the addition of a death animation, as well as code to cause damage to the player. The player's side lacks the writing of a falling death system and a rebirth birth system, as well as causing damage to the Gnome. After these logic codes are completed, the remaining two levels will be completed soon. I expect to complete these logics tomorrow and then complete the second and third levels.